+++ src/components/settings/CartSettingsEditor.tsx
@@ -9,6 +9,7 @@ import { SettingRowInput, SettingRowLabel } from "ui/form/SettingRow";
 import { CartType } from "store/features/settings/settingsState";
 import { Checkbox } from "ui/form/Checkbox";
 import { useAppDispatch, useAppSelector } from "store/hooks";
+import Alert from "ui/alerts/Alert";
 
 export interface CartSettingsEditorProps {
   searchTerm?: string;
@@ -96,6 +97,13 @@ const CartSettingsEditor = ({ searchTerm }: CartSettingsEditorProps) => {
             checked={batterylessEnabled}
             onChange={onToggleBatteryless}
           />
+          {batterylessEnabled && (
+            <Alert variant="warning" style={{ marginTop: 3 }}>
+              <p>{l10n("FIELD_CART_BATTERYLESS_MORE_INFO_1")}</p>
+              <p>{l10n("FIELD_CART_BATTERYLESS_MORE_INFO_2")}</p>
+              <p>{l10n("FIELD_CART_BATTERYLESS_MORE_INFO_3")}</p>
+            </Alert>
+          )}
         </SettingRowInput>
       </SearchableSettingRow>
       {!searchTerm && (