+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated Simplified Chinese localisation. [@wcxu21](https://github.com/wcxu21)
 - Improve legibility of extra collision tiles by using a pixel font "Public Pixel" by GGBot [@kinostl](https://github.com/kinostl)
 - Updated Polish localisation. [@ReptiIe](https://github.com/ReptiIe)
+- Changed wording of batteryless save feature and added warning when feature is enabled to explain that only very specific cartridges will be supported
 
 ### Fixed